backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File Processing
python-multipart==0.0.6
python-slugify==8.0.1
aiofiles==23.2.1

# Data Analysis
pandas==2.1.3
openpyxl==3.1.2

# HTTP Requests
requests==2.31.0
pywebpush==1.14.0

# Environment
python-dotenv==1.0.0

# Templates
jinja2==3.1.2
firebase-admin>=6.0.0

# AI
openai==1.3.0